logging improve

+++ src/main/java/usermanager/controller/UserTypeController.java
@@ -7,47 +7,53 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import usermanager.model.User;
 import usermanager.model.UserType;
+import usermanager.service.UserService;
 import usermanager.service.UserTypeService;
 import usermanager.util.CustomResponse;
 import usermanager.util.UserTypeValidator;
-
 import java.util.HashMap;
-
+import java.util.List;
 
 @Controller
-@RequestMapping(path="/usertypes")
+@RequestMapping(path = "/usertypes")
 public class UserTypeController {
     @Autowired
     private UserTypeService userTypeService;
     @Autowired
     private UserTypeValidator userTypeValidator;
+    @Autowired
+    private UserService userService;
+
     @Bean
     private DozerBeanMapper dozerBeanMapper() {
         DozerBeanMapper mapper = new DozerBeanMapper();
-        mapper.setCustomFieldMapper((source, destination, sourceFieldValue, classMap, fieldMapping) -> sourceFieldValue == null);
+        mapper.setCustomFieldMapper(
+                (source, destination, sourceFieldValue, classMap, fieldMapping) -> sourceFieldValue == null);
         return mapper;
     }
 
-    @RequestMapping(path="/" , method = RequestMethod.GET)
+    @RequestMapping(path = "/", method = RequestMethod.GET)
     public ResponseEntity<?> getAllUserTypes() {
-        return new ResponseEntity<>(new CustomResponse(200,"success",userTypeService.findAll()),HttpStatus.OK);
+        return new ResponseEntity<>(new CustomResponse(200, "success", userTypeService.findAll()), HttpStatus.OK);
     }
 
-    @RequestMapping(path="/{id}" , method = RequestMethod.GET)
+    @RequestMapping(path = "/{id}", method = RequestMethod.GET)
     public ResponseEntity<?> getAllUserTypes(@PathVariable String id) {
         int Id;
-        try{
+        try {
             Id = Integer.parseInt(id);
-        }catch (Exception NumberFormatException){
-            return new ResponseEntity<>(new CustomResponse(400,"invalid data type for ID",null), HttpStatus.BAD_REQUEST);
+        } catch (Exception NumberFormatException) {
+            return new ResponseEntity<>(new CustomResponse(400, "invalid data type for ID", null),
+                    HttpStatus.BAD_REQUEST);
         }
         UserType userType = userTypeService.findOne(Id);
-        if(userType!=null){
-            return new ResponseEntity<>(new CustomResponse(200,"success",userType), HttpStatus.OK);
-        }
-        else {
-            return new ResponseEntity<>(new CustomResponse(400,"user type does not exist",null), HttpStatus.BAD_REQUEST);
+        if (userType != null) {
+            return new ResponseEntity<>(new CustomResponse(200, "success", userType), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(new CustomResponse(400, "user type does not exist", null),
+                    HttpStatus.BAD_REQUEST);
         }
 
     }
@@ -55,54 +61,61 @@ public class UserTypeController {
     @RequestMapping(value = "/", method = RequestMethod.POST)
     public ResponseEntity<?> update(@RequestBody UserType userType) {
         HashMap err = userTypeValidator.validate(userType);
-        if(err.isEmpty()){
+        if (err.isEmpty()) {
             userTypeService.save(userType);
-            return new ResponseEntity<>(new CustomResponse(200,"success",userType),HttpStatus.OK);
-        }
-        else{
-            return new ResponseEntity<>(new CustomResponse(400,err,null),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new CustomResponse(200, "success", userType), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(new CustomResponse(400, err, null), HttpStatus.BAD_REQUEST);
         }
     }
 
-    @RequestMapping(path="/{id}" , method = RequestMethod.DELETE)
+    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteOne(@PathVariable String id) {
         int Id;
-        try{
+        try {
             Id = Integer.parseInt(id);
-        }catch (Exception NumberFormatException){
-            return new ResponseEntity<>(new CustomResponse(400,"invalid data type for ID",null), HttpStatus.BAD_REQUEST);
+        } catch (Exception NumberFormatException) {
+            return new ResponseEntity<>(new CustomResponse(400, "invalid data type for ID", null),
+                    HttpStatus.BAD_REQUEST);
         }
         UserType userType = userTypeService.findOne(Id);
-        if(userType!=null){
-            userTypeService.deleteOne(Id);
-            return new ResponseEntity<>(new CustomResponse(200,"success",userType), HttpStatus.OK);
-        }
-        else {
-            return new ResponseEntity<>(new CustomResponse(400,"user type does not exist",null), HttpStatus.BAD_REQUEST);
+        if (userType != null) {
+            List<User> assignUsers = userService.findAllByUserType(userType);
+            if (assignUsers.size() > 0) {
+                return new ResponseEntity<>(new CustomResponse(400, "user type does is assined to users", null),
+                        HttpStatus.BAD_REQUEST);
+            } else {
+                userTypeService.deleteOne(Id);
+                return new ResponseEntity<>(new CustomResponse(200, "success", userType), HttpStatus.OK);
+            }
+        } else {
+            return new ResponseEntity<>(new CustomResponse(400, "user type does not exist", null),
+                    HttpStatus.BAD_REQUEST);
         }
 
     }
 
-    @RequestMapping(path="/{id}" , method = RequestMethod.PUT)
+    @RequestMapping(path = "/{id}", method = RequestMethod.PUT)
     public ResponseEntity<?> updateOne(@RequestBody UserType userType, @PathVariable String id) {
         int Id;
-        try{
+        try {
             Id = Integer.parseInt(id);
-        }catch (Exception NumberFormatException){
-            return new ResponseEntity<>(new CustomResponse(400,"invalid data type for ID",null), HttpStatus.BAD_REQUEST);
+        } catch (Exception NumberFormatException) {
+            return new ResponseEntity<>(new CustomResponse(400, "invalid data type for ID", null),
+                    HttpStatus.BAD_REQUEST);
         }
         UserType oldUserType = userTypeService.findOne(Id);
-        if(oldUserType!=null){
-            dozerBeanMapper().map(userType,oldUserType);
+        if (oldUserType != null) {
+            dozerBeanMapper().map(userType, oldUserType);
             HashMap err = userTypeValidator.validate(oldUserType);
-            if(err.isEmpty()){
+            if (err.isEmpty()) {
                 UserType updatedUserType = userTypeService.update(oldUserType);
-                return new ResponseEntity<>(new CustomResponse(200,"success",updatedUserType), HttpStatus.OK);
+                return new ResponseEntity<>(new CustomResponse(200, "success", updatedUserType), HttpStatus.OK);
             }
-            return new ResponseEntity<>(new CustomResponse(400,err ,null), HttpStatus.BAD_REQUEST);
-        }
-        else {
-            return new ResponseEntity<>(new CustomResponse(400,"user type does not exist",null), HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new CustomResponse(400, err, null), HttpStatus.BAD_REQUEST);
+        } else {
+            return new ResponseEntity<>(new CustomResponse(400, "user type does not exist", null),
+                    HttpStatus.BAD_REQUEST);
         }
 
     }
